All Code and images for TinDog
TinDog, a fictional site, is a play on the dating app Tinder- except this time, it's for dogs. This was an instructional module on building a simple webpage, utilizing Twitter's BootStrap framework. Much of the code was independently written, without instruction from Dr. Yu, in the form of solo-implementation challenges. CSS refactoring was also a main point of the assignment, with heavy emphasis on clean, legible code.